Move athletes POST methods to API
