+++ app/views/athletes/_payment.html.erb
@@ -1,4 +1,4 @@
-<%= form_tag({controller: 'athletes', action: 'subscribe_to_pro'}, id: 'form-subscribe-to-pro') do %>
+<%= form_tag({controller: 'api/athletes', action: 'subscribe_to_pro'}, id: 'form-subscribe-to-pro') do %>
   <input type="hidden" name="subscriptionPlanId" />
   <input type="hidden" name="stripeToken" />
   <input type="hidden" name="stripeEmail" />
