+++ config/routes.rb
@@ -10,11 +10,6 @@ Rails.application.routes.draw do
   get 'athletes/:id/get-pro' => 'athletes#pro_plans'
   get 'athletes/:id/cancel-pro' => 'athletes#cancel_pro'
 
-  post 'athletes/:id/fetch_latest' => 'athletes#fetch_latest'
-  post 'athletes/:id/save_profile' => 'athletes#save_profile'
-  post 'athletes/:id/reset_profile' => 'athletes#reset_profile'
-  post 'athletes/:id/subscribe_to_pro' => 'athletes#subscribe_to_pro'
-
   namespace :api do
     get 'athletes/:id/meta' => 'meta#index'
     get 'athletes/:id/best-efforts' => 'best_efforts#index'
@@ -24,6 +19,12 @@ Rails.application.routes.draw do
     get 'athletes/:id/personal-bests/:distance' => 'personal_bests#index'
     get 'athletes/:id/races' => 'races#index'
     get 'athletes/:id/races/:distance_or_year' => 'races#index'
+
+    post 'athletes/:id/fetch_latest' => 'athletes#fetch_latest'
+    post 'athletes/:id/save_profile' => 'athletes#save_profile'
+    post 'athletes/:id/reset_profile' => 'athletes#reset_profile'
+    post 'athletes/:id/subscribe_to_pro' => 'athletes#subscribe_to_pro'
+
     get 'faqs/index' => 'faqs#index'
   end
