+++ spec/requests/api/athletes_fetch_latest_spec.rb
@@ -1,15 +1,15 @@
 require 'rails_helper'
 
-RSpec.describe AthletesController, type: :request do
+RSpec.describe Api::AthletesController, type: :request do
   describe 'POST fetch_latest' do
     it 'should return 404 when the requested athlete does not exist' do
-      post '/athletes/12345678/fetch_latest'
+      post '/api/athletes/12345678/fetch_latest'
       expect(response).to have_http_status(404)
     end
 
     it 'should return 403 when requested athlete is not the current user' do
       setup_cookie(nil)
-      post '/athletes/9123806/fetch_latest'
+      post '/api/athletes/9123806/fetch_latest'
       expect(response).to have_http_status(403)
     end
 
@@ -19,7 +19,7 @@ RSpec.describe AthletesController, type: :request do
         setup_cookie('4d5cf2bbc714a4e22e309cf5fcf15e40')
 
         # act.
-        post '/athletes/9123806/fetch_latest'
+        post '/api/athletes/9123806/fetch_latest'
 
         # assert.
         expect(response).to have_http_status(:success)
@@ -32,7 +32,7 @@ RSpec.describe AthletesController, type: :request do
         setup_cookie('58e42e6f5e496dc5aa0d5ec354da8048')
 
         # act.
-        post '/athletes/456/fetch_latest'
+        post '/api/athletes/456/fetch_latest'
 
         # assert.
         expect(response).to have_http_status(403)
