+++ spec/requests/api/meta_api_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe Api::MetaController, type: :request do
   describe 'GET index' do
     it 'should not find athlete by id who does not exist' do
       expect { get '/api/athletes/12345678/meta' }
-        .to raise_error(ActionController::RoutingError, "Could not find athlete '12345678' by id.")
+        .to raise_error(ActionController::RoutingError, "Could not find the requested athlete '12345678' by id.")
     end
 
     it 'should be successful for an existing athlete' do
