+++ spec/requests/api/races_api_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe Api::RacesController, type: :request do
   describe 'GET index' do
     it 'should not find athlete by id who does not exist' do
       expect { get '/api/athletes/12345678/races' }
-        .to raise_error(ActionController::RoutingError, "Could not find athlete '12345678' by id.")
+        .to raise_error(ActionController::RoutingError, "Could not find the requested athlete '12345678' by id.")
     end
 
     it 'should be a 404 with an invalid distance' do
