+++ spec/requests/expected/athletes/9123806/get-pro.html
@@ -446,7 +446,7 @@ height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
     </section>
   </div><!-- /.content-wrapper -->
 
-  <form id="form-subscribe-to-pro" action="/athletes/9123806/subscribe_to_pro" accept-charset="UTF-8" method="post"><input name="utf8" type="hidden" value="&#x2713;" />
+  <form id="form-subscribe-to-pro" action="/api/athletes/9123806/subscribe_to_pro" accept-charset="UTF-8" method="post"><input name="utf8" type="hidden" value="&#x2713;" />
   <input type="hidden" name="subscriptionPlanId" />
   <input type="hidden" name="stripeToken" />
   <input type="hidden" name="stripeEmail" />
